// frontend/src/app/api/model/dashboard.ts
/**
 * QuickFlow API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Habit } from './habit';
import { Task } from './task';
import { PlanCounts } from './planCounts';
import { HabitCounts } from './habitCounts';
import { TaskCounts } from './taskCounts';
import { Plan } from './plan';
import { LearningSnapshot } from './learningSnapshot';


export interface Dashboard { 
    now: string;
    dueToday: Array<Task>;
    overdue: Array<Task>;
    completedToday: Array<Task>;
    taskCompletionPercent: number;
    taskCounts: TaskCounts;
    /**
     * Active habits
     */
    habits: Array<Habit>;
    habitCounts: HabitCounts;
    /**
     * IN_PROGRESS plans by priority
     */
    activePlans: Array<Plan>;
    planCounts: PlanCounts;
    learning: LearningSnapshot;
}