// frontend/src/app/api/model/planItem.ts
/**
 * QuickFlow API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PlanSourceType } from './planSourceType';


export interface PlanItem { 
    id: number;
    sourceType: PlanSourceType;
    sourceId: number;
    sourceTitle: string;
    /**
     * The source was deleted (FR-07.6)
     */
    sourceRemoved: boolean;
    done: boolean;
}
export namespace PlanItem {
}